refactor: 이동 전략 템플릿 메서드 패턴 적용
